Fix errors on Android from moving FontLoader.js
